Segregating now_variable() et al from executing.py
